// orto-protesi/src/app/models/cliente.model.ts
export interface Cliente {
  id?: number;
  nomeDottore: string;
  cognomeDottore: string;
  email: string;
  telefonoCellulare: string;
  telefonoFisso: string;
  codiceFiscale: string;
  partitaIva: string;
  nomeStudio: string;
  viaStudio: string;
  cittaStudio: string;
  codiceSDI: string;
  dataCreazione: Date;
  
}